[FP-32] 프로필 최소 금리 자료구조 변경

+++ finpik-domain/profile/src/main/java/finpik/entity/Profile.java
@@ -34,7 +34,7 @@ public class Profile {
     private OccupationDetail occupationDetail;
     private User user;
     private Integer recommendedLoanProductCount;
-    private Double minInterestRate;
+    private Float minInterestRate;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
